requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pyodbc==5.1.0
python-dotenv==1.0.0
pydantic-settings==2.10.1
structlog==23.2.0
anthropic>=0.28.0
python-multipart>=0.0.9
pytest==7.4.3
httpx==0.27.0
